+++ src/main/java/net/soulsweaponry/items/abilities/abilitykeybind/BasicKeybindAbility.java
@@ -6,6 +6,7 @@ import net.minecraft.item.ItemStack;
 import net.minecraft.server.world.ServerWorld;
 import net.minecraft.text.Text;
 import net.minecraft.util.Hand;
+import net.minecraft.world.World;
 import net.soulsweaponry.util.WeaponUtil;
 import org.apache.logging.log4j.util.TriConsumer;
 import org.jetbrains.annotations.Nullable;
@@ -16,7 +17,7 @@ import java.util.List;
  * Template for a simple keybind ability such as just adding an effect (server side) and playing a sound effect (client side).
  * Will not trigger if the player has {@link net.soulsweaponry.registry.EffectRegistry#COOLDOWN} effect.
  * @param serverEffects server side effects, such as spawning entities
- * @param clientEffects client side effects, such as particles or sounds
+ * @param clientEffects client side effects, such as particles or sounds, the world is of a client world
  * @param tooltips tooltips of the ability
  * @param stackDamage
  * @param minCooldown
@@ -25,7 +26,7 @@ import java.util.List;
  */
 public record BasicKeybindAbility(
         TriConsumer<ServerWorld, ItemStack, PlayerEntity> serverEffects,
-        TriConsumer<ClientWorld, ItemStack, PlayerEntity> clientEffects,
+        TriConsumer<World, ItemStack, PlayerEntity> clientEffects,
         List<Text> tooltips, int stackDamage, int minCooldown, int cooldown, int reducedCooldownPerLvl
 ) implements IKeybindAbility {